Add ResourceSet support
The ResourceSet is very useful in case there is many metamodel to deal
with. It creates separated resources where each produced metamodel is
stored. It also eases the metamodel serialization. This addition is
the first step towards a 'multiple' metamodel serialization at once
(but other issues, like the non-generation of some already generated
metamodel should be addressed).