Add special '__name__' assignement for new classes
In many part of textX, classes name are tested using the '__name__'
property. This makes sences as each type instance is a class in
textX. However, with PyEcore, elements can be EClasses (which are kind
of classes), EEnum and EDatatype. The two latter types are not class
per say, but instances of classes. If the '__name__' property is not
set on these instances, all the current textX code fails. Another
solution would have been to change all occurences of 'xxx.__name__' to
'xxx.name', but this modification would have been put the fork far
away from the original branch.