Add fake resource with the fake EPackage for basic EDatatype
This fake resource is useful if some serialization is required. It has
no real interest as EDatatype are used, but if EClasses are inserted
in the fake EPackage, it would be necessary. As long as the fake
resource is not inserted in the resource set, there will be no issues.