Add basic textX types to Ecore instead of a fake metamodel
If added to a fake metamodel, the XMI serialization for a fake
metamodel is problematic. As the fake metamodel is not registered
under a metamodel registry, the XMI serializer does not find the good
URI which is needed for the serialization. Instead, this commit adds
the basic textX elements in the Ecore metamodel. This way, the
object uri path can be found for serialization.