Add support for tmp meta-attributes
This commit provides a way of dealing with tmp meta-attributes. By tmp
meta-attribute, should be understood: attributes that are expressed in
the grammar, but which are not defined in the metamodel (user
classes). These attributes are threated differently and are gathered
as EClass non EStructuralFeature, but as pure python attribute. This
distinction is important as it ensure that during model serialization
(XMI or JSON), the serialized data will not contain unkown attributes.