Change 'packages' keyword for 'package' in TextXMetamodel
The use of multiple generated PyEcore packages as input of the grammar
parser can be tricky as it suppose that the main EPackage could be any
of those. This is not an issue in itself until XMI serialization is
performed. In this case, it become complex to know which EPackage
should be the default package/namespace. This restriction removes this
issue but could introduce difficulties for XMI serialization when a
metamodel hierarchie is used.