Removes the 'parent' keyword when it is not declared in the __init__
The code generated by pyecoregen imposes that any parameter used in a
constructor must be consummed somewhere in the object
hierarchie. However, in TextX, the 'parent' keyword is systematically
added to the object constructor parameters, and it raises an error
when used with code generated by pyecoregen. To avoid that, the code
now checks if the 'parent' parameter is explicitally used in the
constructor, if it is the case, the parameter is not removed. This
solution is not ideal, but, at least, let the user the choice to use
the 'parent' keyword or not, without changing the pyecoregen
generator.